[kernel] usbstorage.c: New function USBStorage_GetSectorInfo().
This function returns the sector size and count of the mounted
USB storage device. It's used by disk_initialize() to set up
FF_dev_info[].

This eliminates usbstorage.c's dependency on diskio.h.

[fatfs] integer.h: Use int for LONG/DWORD instead of long, since
gcc keeps complaining about mixing it with u32.

+++ fatfs/integer.h
@@ -15,20 +15,20 @@ typedef unsigned __int64 QWORD;
 #else			/* Embedded platform */
 
 /* These types MUST be 16-bit or 32-bit */
-typedef int				INT;
+typedef int		INT;
 typedef unsigned int	UINT;
 
 /* This type MUST be 8-bit */
 typedef unsigned char	BYTE;
 
 /* These types MUST be 16-bit */
-typedef short			SHORT;
+typedef short		SHORT;
 typedef unsigned short	WORD;
 typedef unsigned short	WCHAR;
 
 /* These types MUST be 32-bit */
-typedef long			LONG;
-typedef unsigned long	DWORD;
+typedef int		LONG;
+typedef unsigned int	DWORD;
 
 /* This type MUST be 64-bit (Remove this for C89 compatibility) */
 typedef unsigned long long QWORD;
